query-module/requirements.txt
# FastAPI и ASGI сервер
fastapi==0.115.5
uvicorn[standard]==0.34.0

# HTTP клиент для взаимодействия с Storage Element
httpx==0.28.1

# Database (async PostgreSQL)
asyncpg==0.30.0
SQLAlchemy==2.0.36
alembic==1.14.0

# Redis для кеширования (SYNC режим согласно архитектуре)
redis==5.2.1

# JWT authentication (RS256)
PyJWT[crypto]==2.10.1
cryptography==44.0.0

# Pydantic для валидации
pydantic==2.10.3
pydantic-settings==2.6.1

# Structured logging
python-json-logger==3.2.1

# Prometheus metrics
prometheus-client==0.21.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-timeout==2.3.1
httpx-mock==0.1.1

# Code quality
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.13.0

# Development
python-dotenv==1.0.1
